+++ packages/react/src/components/SheetOverlay/drag_and_drop/column-helpers.tsx
@@ -254,7 +254,6 @@ export const useColumnDragAndDrop = (
         Number.isFinite(finalInsertionIndex) &&
         finalInsertionIndex >= 0
       ) {
-
         setContext((draft) => {
           const _sheet = draft.luckysheetfile[sheetIdx];
           if (!_sheet?.data) return;
@@ -281,6 +280,8 @@ export const useColumnDragAndDrop = (
 
           _sheet.data = rows;
           updateContextWithSheetData(draft, _sheet.data);
+
+          // update formula
           const d = getFlowdata(draft);
           d?.forEach((row) => {
             row.forEach((cell) => {
@@ -364,6 +365,28 @@ export const useColumnDragAndDrop = (
             });
           });
 
+          // update dataVerification
+          if (_sheet.dataVerification) {
+          const newDataVerification: any = {};
+          Object.keys(_sheet.dataVerification).forEach((item) => {
+            const itemData = _sheet.dataVerification?.[item];
+            const colRow = item.split("_");
+            if (colRow.length !== 2) return;
+            const presentcol = parseInt(colRow[1], 10);
+            let updatedCol = presentcol;
+            if (presentcol === sourceIndex) {
+              updatedCol = targetIndex;
+            } else if (presentcol > sourceIndex && presentcol < targetIndex) {
+              updatedCol -= 1;
+            } else if (presentcol < sourceIndex && presentcol >= targetIndex) {
+              updatedCol += 1;
+            }
+            newDataVerification[`${colRow[0]}_${updatedCol}`] = itemData;
+          });
+          _sheet.dataVerification = newDataVerification;
+        }
+
+          // update calc chain
           _sheet.calcChain?.forEach((item) => {
             if (item.c === sourceIndex) {
               item.c = targetIndex;
@@ -373,6 +396,8 @@ export const useColumnDragAndDrop = (
               item.c += 1;
             }
           });
+
+          // update data block
           // @ts-expect-error
           window?.updateDataBlockCalcFunctionAfterRowDrag?.(
             sourceIndex,