+++ packages/react/src/components/Toolbar/index.tsx
@@ -37,6 +37,10 @@ import {
   insertDuneChart,
   Cell,
 } from "@fileverse-dev/fortune-core";
+import {
+  setSelection,
+  getSelection,
+} from "@fileverse-dev/fortune-core/src/api";
 import _ from "lodash";
 import {
   IconButton,
@@ -49,6 +53,8 @@ import {
   CommandGroup,
   CommandItem,
 } from "@fileverse/ui";
+import DataVerificationPortal from "./dataVerificationPortal";
+// import DataVerification from "../DataVerification";
 import WorkbookContext from "../../context";
 import "./index.css";
 import Button from "./Button";
@@ -59,7 +65,6 @@ import SVGIcon from "../SVGIcon";
 import { useDialog } from "../../hooks/useDialog";
 import { SplitColumn } from "../SplitColumn";
 import { LocationCondition } from "../LocationCondition";
-import DataVerification from "../DataVerification";
 import ConditionalFormat from "../ConditionFormat";
 import CustomButton from "./CustomButton";
 import { CustomColor } from "./CustomColor";
@@ -608,6 +613,8 @@ const Toolbar: React.FC<{
     isDesktop,
   ]);
 
+  const [showDataValidation, setShowDataValidation] = useState(false);
+
   const getToolbarItem = useCallback(
     (name: string, i: number) => {
       // @ts-ignore
@@ -1080,19 +1087,34 @@ const Toolbar: React.FC<{
       }
       if (name === "dataVerification") {
         return (
-          <Button
-            iconId={name}
-            tooltip={tooltip}
-            key={name}
-            onClick={() => {
-              if (context.allowEdit === false) return;
-              showDialog(
-                <DataVerification />,
-                undefined,
-                toolbar.dataVerification
-              );
-            }}
-          />
+          <>
+            <DataVerificationPortal visible={showDataValidation} />
+            <Button
+              iconId={name}
+              tooltip={tooltip}
+              key={name}
+              onClick={() => {
+                const selection = getSelection(context);
+                if (!selection) {
+                  setContext((ctx) => {
+                    setSelection(ctx, [{ row: [0, 0], column: [0, 0] }], {
+                      id: context.currentSheetId,
+                    });
+                  });
+                }
+                document.getElementById("data-verification-button")?.click();
+                // if (context.allowEdit === false) return;
+                // showDialog(
+                //   <DataVerification />,
+                //   undefined,
+                //   toolbar.dataVerification
+                // );
+                setTimeout(() => {
+                  setShowDataValidation(true);
+                }, 100);
+              }}
+            />
+          </>
         );
       }
       if (name === "locationCondition") {
@@ -1970,6 +1992,7 @@ const Toolbar: React.FC<{
       customColor,
       customStyle,
       toolbarFormat.moreCurrency,
+      context.dataVerification?.dataRegulation,
     ]
   );